requirements.txt
haystack-ai
pypdf
streamlit
dash
dash_mantine_components
dash_iconify
python-dotenv